+++ src/submissions/frontend/visualizations/__init__.py
@@ -2,3 +2,4 @@
 Contains all operations for creating charts, graphs and visual effects.
 '''
 from .control_charts import *
+from .barcode import *